backend/requirements.txt
Flask==3.0.3
flask-cors==5.0.0
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
watchdog==5.0.3
PyMuPDF==1.24.12
pandas==2.2.3
numpy==2.0.2
pymupdf4llm
xlsxwriter
